+++ pages/src/components/Tooltip/Tooltip.js
@@ -6,11 +6,35 @@ var Tooltip = Backbone.View.extend({
 
         var view = this;
         $.get('src/components/Tooltip/tooltip.html', function(tpl) {
-            view.$el.html(Mustache.render(tpl));
+            view.$el.html(Mustache.render(tpl));view.$el.html(Mustache.render(tpl));
+			view.initBizUI();
+			hljs.initHighlighting.called = false;
+			hljs.initHighlighting();
         });
     },
 
+	initBizUI: function() {
+        $('input').bizInput();
+        $('button').bizButton();
+        bizui.Tooltip({
+            element: 'input',
+            theme: 'orange'
+        });
+        bizui.Tooltip({
+            element: 'button',
+            action: 'hover'
+        });
+        bizui.Tooltip({
+            element: '.protocol',
+            preventDefault: true,
+            action: 'click'
+        });
+	},
+
     destroy: function() {
+        bizui.Tooltip({
+            removeAll: true
+        });
         this.$el.empty();
     }
 });